♻️ refactor: Light FSD 기반 폴더 구조 전면 재편
- shared/{ui,lib,assets} 레이어 신설: Button, ImageUploader, axios, utils, icon 이동
- features/project-register → features/project/register (확장 대비 도메인 그룹화)
- features/project/register 내부 FSD 구조 적용: ui/, model/, index.ts(public API barrel)
- components/sidebar, components/toast 전역 위젯 분리 및 sidebar 내부 menu/dropdown 논리 그룹화
- useToastStore를 components/toast 모듈로 통합 (stores/ 폴더 제거)
- 빈 placeholder 폴더 제거 (api, hooks, types, stores, components/ui, components/layout)
- routes/__root.tsx, matching/route.tsx, test/* import 경로 업데이트
- package.json svgr 스크립트 경로 업데이트
- README 프로젝트 구조 섹션 갱신
