--- src/components/common/sidebar/dropdown/SideBarDropDownMenuItem.tsx
@@ -1,27 +1,0 @@
-import CheckIcon from "@/assets/icon/check/CheckIcon"
-import { cn } from "@/lib/utils"
-
-interface SideBarDropDownMenuItemProps {
-  label: string
-  isSelected: boolean
-  onClick: () => void
-}
-
-export function SideBarDropDownMenuItem({
-  label,
-  isSelected,
-  onClick,
-}: SideBarDropDownMenuItemProps) {
-  return (
-    <button
-      onClick={onClick}
-      className={cn(
-        "text-body-3-regular flex h-6.5 w-42 items-center justify-between rounded-[8px] py-1 pr-2 pl-3 text-left",
-        isSelected ? "bg-teal-50 text-teal-600" : "text-teal-gray-600",
-      )}
-    >
-      <span>{label}</span>
-      {isSelected && <CheckIcon className="h-4 w-4" />}
-    </button>
-  )
-}
