+++ src/components/sidebar/SideBar.tsx
@@ -1,0 +1,52 @@
+import { useState } from "react"
+
+import { cn } from "@/shared/lib/utils"
+
+import { SideBarDropDown } from "./dropdown/SideBarDropDown"
+import { SideBarMenu } from "./menu/SideBarMenu"
+import { SideBarMenuItem } from "./menu/SideBarMenuItem"
+import { SIDEBAR_ITEMS } from "./sidebar.config"
+
+interface SideBarProps {
+  className?: string
+}
+
+const DEMO_DAY_EDITION = 10
+
+export default function SideBar({ className }: SideBarProps) {
+  const [openTitle, setOpenTitle] = useState<string>(
+    SIDEBAR_ITEMS[0]?.title ?? "",
+  )
+
+  return (
+    <nav
+      aria-label="사이드 메뉴"
+      className={cn(
+        "border-teal-gray-200 flex w-55 shrink-0 flex-col items-center justify-start border-r pt-4",
+        className,
+      )}
+    >
+      <SideBarDropDown />
+      <div className="flex flex-col py-4">
+        <span className="text-body-3-regular text-teal-gray-400 mb-2 pl-0.5">
+          {DEMO_DAY_EDITION}th Demoday
+        </span>
+        {SIDEBAR_ITEMS.map(({ title, icon, menus }) => (
+          <SideBarMenu
+            key={title}
+            title={title}
+            icon={icon}
+            isOpen={openTitle === title}
+            onToggle={() =>
+              setOpenTitle((prev) => (prev === title ? "" : title))
+            }
+          >
+            {menus.map((menu) => (
+              <SideBarMenuItem key={menu.to} title={menu.title} to={menu.to} />
+            ))}
+          </SideBarMenu>
+        ))}
+      </div>
+    </nav>
+  )
+}
