+++ src/features/project/register/model/basicInfoSchema.ts
