+++ src/features/project/register/model/useProjectRegisterStore.ts
@@ -1,0 +1,13 @@
+import { create } from "zustand"
+
+import type { BasicInfoFormData } from "./basicInfoSchema"
+
+interface ProjectRegisterState {
+  basicInfo: BasicInfoFormData | null
+  setBasicInfo: (data: BasicInfoFormData) => void
+}
+
+export const useProjectRegisterStore = create<ProjectRegisterState>((set) => ({
+  basicInfo: null,
+  setBasicInfo: (data) => set({ basicInfo: data }),
+}))
