+++ src/features/project/register/ui/SectionHeader.tsx
