+++ src/features/project/register/ui/Stepper.tsx
