+++ src/shared/assets/icon/check/CheckIcon.tsx
