+++ src/shared/assets/icon/close/CloseIcon.tsx
