+++ src/shared/assets/icon/people/PersonButtonIcon.tsx
