+++ src/shared/assets/icon/people/TeamIcon.tsx
