+++ src/shared/assets/icon/upload/UploadImageIcon.tsx
